refactor: delete deprecated handlers/tool-schemas shim, inline imports
Remove the 9-line re-export shim src/handlers/tool-schemas.ts and update all
three call sites (src/handlers/tools.ts, test/tool-schemas.spec.ts,
test/compare-baseline.spec.ts) to import directly from src/schemas/tool-definitions.
