--- src/handlers/tool-schemas.ts
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: BUSL-1.1
-
-/**
- * MCP Tool schema definitions — re-exported from Zod-derived schemas.
- * @deprecated Import directly from '../schemas/tool-definitions' in new code.
- */
-
-export { TOOLS, TOOL_SCHEMA_MAP } from '../schemas/tool-definitions';
-export type { McpTool, ToolGroup, ToolTier } from '../schemas/tool-definitions';
